requirements.txt
Django==4.2.5
psycopg2==2.9.7
Pillow==10.0.0

